+++ src/main/java/site/coach_coach/coach_coach_server/config/SentryConfiguration.java
@@ -1,10 +1,10 @@
-package site.coach_coach.coach_coach_server.common.exception;
+package site.coach_coach.coach_coach_server.config;
 
 import org.springframework.context.annotation.Configuration;
 
 import io.sentry.spring.jakarta.EnableSentry;
 
-@EnableSentry
+@EnableSentry(dsn = "${sentry.dsn}")
 @Configuration
 class SentryConfiguration {
 }
